Add files via upload
The stress driven dynamics of Notch-Wnt-p53 cross-talk is subjected to a few possible dynamical states governed by simple fractal rules, and allowed to decide its own fate by choosing one of these states which are contributed from long range correlation with varied fluctuations due to active molecular interaction. The topological properties of the networks corresponding to these dynamical states have hierarchical features with assortive structure. The stress signal driven by nutlin and modulated by mediator GSK3 acts as anti-apoptotic signal in this system, whereas, the stress signal driven by Axin and modulated by GSK3 behaves as anti-apoptotic for a certain range of Axin and GSK3 interaction, and beyond which the signal acts as favor-apoptotic signal. However, this stress system prefers to stay in an active dynamical state whose counterpart complex network is closest to hierarchical topology with exhibited roles of few interacting hubs. During the propagation of stress signal, the system allows the propagator pathway to inherit all possible properties of the state to the receiver pathway/pathways with slight modifications, indicating efficient information processing and democratic sharing of responsibilities in the system via cross-talk. The increase in the number of cross-talk pathways in the system favors to establish self-organization.